+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-08-14 15:28:07
-// Commit: feat(network): integrate custom telemetry logger in network
+// Updated at: 2026-08-14 18:08:16
+// Commit: fix(network): fix memory leak by cancelling network subscription
 
 class networkComponentState {
-  final String commitRef = '3609';
-  final String lastMessage = 'feat(network): integrate custom telemetry logger in network';
-  final DateTime timestamp = DateTime.parse('2026-08-14T15:28:07+05:00');
+  final String commitRef = '3617';
+  final String lastMessage = 'fix(network): fix memory leak by cancelling network subscription';
+  final DateTime timestamp = DateTime.parse('2026-08-14T18:08:16+05:00');
 }